+++ grid_esp/main/grid_esp32_module_pbf4.c
@@ -56,7 +56,7 @@ void grid_esp32_module_pbf4_task(void *arg)
     static uint8_t multiplexer_index =0;
     static const uint8_t multiplexer_lookup[16] = {0, 1, 4, 5, 8, 9, 12, 13, 2, 3, 6, 7, 10, 11, 14, 15};
 
-    static uint32_t last_real_time[16] = {0};
+    static uint64_t last_real_time[16] = {0};
